+++ web/components/core/modals/index.ts
@@ -4,3 +4,4 @@ export * from "./gpt-assistant-popover";
 export * from "./link-modal";
 export * from "./user-image-upload-modal";
 export * from "./workspace-image-upload-modal";
+export * from "./issue-search-modal-empty-state";
