+++ web/components/issues/parent-issues-list-modal.tsx
@@ -8,15 +8,13 @@ import { ProjectService } from "services/project";
 import useDebounce from "hooks/use-debounce";
 import { usePlatformOS } from "hooks/use-platform-os";
 // components
-import { EmptyState } from "components/empty-state";
+import { IssueSearchModalEmptyState } from "components/core";
 // ui
 import { Loader, ToggleSwitch, Tooltip } from "@plane/ui";
 // icons
 import { Rocket, Search } from "lucide-react";
 // types
 import { ISearchIssueResponse } from "@plane/types";
-// constants
-import { EmptyStateType } from "constants/empty-state";
 
 type Props = {
   isOpen: boolean;
@@ -69,19 +67,6 @@ export const ParentIssuesListModal: React.FC<Props> = ({
       .then((res) => setIssues(res))
       .finally(() => setIsSearching(false));
   }, [debouncedSearchTerm, isOpen, issueId, isWorkspaceLevel, projectId, workspaceSlug]);
-
-  const renderEmptyState = (type: EmptyStateType) => (
-    <div className="flex flex-col items-center justify-center px-3 py-8 text-center">
-      <EmptyState type={type} layout="screen-simple" />
-    </div>
-  );
-
-  const emptyState =
-    issues.length === 0 && searchTerm !== "" && debouncedSearchTerm !== "" && !isSearching
-      ? renderEmptyState(EmptyStateType.ISSUE_RELATION_SEARCH_EMPTY_STATE)
-      : issues.length === 0
-      ? renderEmptyState(EmptyStateType.ISSUE_RELATION_EMPTY_STATE)
-      : null;
 
   return (
     <>
@@ -168,7 +153,12 @@ export const ParentIssuesListModal: React.FC<Props> = ({
                       </h5>
                     )}
 
-                    {emptyState}
+                    <IssueSearchModalEmptyState
+                      debouncedSearchTerm={debouncedSearchTerm}
+                      isSearching={isSearching}
+                      issues={issues}
+                      searchTerm={searchTerm}
+                    />
 
                     {isSearching ? (
                       <Loader className="space-y-3 p-3">
